Fix for 64-bit OS.

+++ UMTS/IntegrityProtect.cpp
@@ -26,14 +26,14 @@
 
 typedef unsigned char u8;
 typedef unsigned short u16;
-typedef unsigned long u32;
+typedef unsigned int u32;
 
 /*---------------------------------------------------------
  *                   Kasumi.h
  *---------------------------------------------------------*/
 typedef unsigned char     u8;
 typedef unsigned short   u16;
-typedef unsigned long    u32;
+typedef unsigned int    u32;
 //void KeySchedule( u8 *key );
 //void Kasumi( u8 *data );
